FailureHandler was failing to handle failures
Since we started wrapping the handler state to enable encryption, the error entity added to the input json by aws is now a level above the handler state so not available to the handler. Since we were only using it for logging anyway I've just removed the logging that was causing the problem.