Refactor abtest.js to automatically submit A/B impressions to Ophan
This required a little reworking of some of the A/B test types, mainly
removing the hardcoded enum and instead using `$Keys<T>`
(https://flow.org/en/docs/types/utilities/) from the A/B
definition (which is now an object keyed by test ID rather than an
array).

The result of this is hopefully a simpler process for adding A/B tests:
just add them to abtests.js and they'll be fired to Ophan automatically.