Updated this app to support a US-based Stripe account (#175)
Updated this app to support a US-based Stripe account. It now determines which Stripe account to use based on the public key being sent along too in the request. For backwards compatibility, that field is optional, and when it's not there the API will use the prior behaviour of choosing the AU-based Stripe account when the currency is AUD, and the UK-based account for all other requests.