feat: added GW regional copy
On the 17 February, the Awareness team are activating a marketing campaign localised to Berlin, during the same time, Guardian Weekly will be refreshing much of it's off-platform marketing. To align with both strands of activity and maintain a coherent customer journey, the headlines on regional landing pages for GW have been updated.