Phase 3 of the cdk migration
In this phase we clean up all our old infrastructure. This will leave us
with a single set of resources, but now being specified with cdk constructs.
This involves dropping the `CfnInclude` construct, deleting the old cloudformation
files, and cleaning up the temporary changes we needed whilst having a dual stack.

To deploy this we need to:

1. Wait for phase 2 to be deployed
2. Deploy the cloudformation manually (It will delete the old infrastructure which riff-raff will block)
3. Merge PR

And then tada! We've finished migrating our infrastructure to cdk!