Phase 2 of cdk migration
In this phase we use cdk to create a 'dual stack' of all of our infrastructure.
We retain the `CfnInclude` construct to directly import our original
cloudformation template but also recreate all of our resources directly
with cdk (mainly with the GuEc2App construct). This means we will have a second
load balancer and autoscaling group etc. We can deploy this to production
and then, when we're happy, we can point fastly at the new load balancer to send
traffic to the new stack. In a follow up PR, we can drop `CfnInclude` and delete
our old cloudformation template.

To deploy this we need to:

1. Manually deploy the cloudformation change (it adds new parameters so riffraff will fail)
2. Merge PR and wait for build and deploy
3. Point fastly at the new load balancer