Use default 5xx alarms provided by the GuEc2App pattern
Currently we have two custom 5xx alarms defined: one for the
load balancer; and one for the instances. This change replaces
both of those with a single alarm provided to us by the
GuEc2App pattern. This alarm looks at the total percentage
of 5xx responses:

((5xxs from load balancer + 5xxs from instances) / total number of requests)

The benefits of switching to this alarm are:

  - reduction of boilerplate in this stack definition
  - picking up improvements 'for free' if/when DevX establish best practices for monitoring/alerting in this area
  - standardising this alarm across all of your apps more easily