Update cmp library and remove dynamic imports.
Previously we were importing the cmp library with dynamic imports.
This was to avoid loading the cmp during SSR which lead to build errors.
Those errors are no longer an issue so we removed the dynamic imports
and replace them with regular imports. We also updated the version of the
cmp library we were using as we were quite a few major versions behind.