Fix the ts errrors in `StripeCardFormContainer`
This required updating to the latest version (1.8.0) of the `react-stripe-js`
package as it required a type update from the package itself to fix. The issue
is that react no longer adds the `children` prop to function components, which
broke the types coming from `react-stripe-js`. They fixed it in 1.7.2, but the
change log since then has very little in it so it seemed to make sense to upgrade
all the way to the latest version.