Increase 5xx alarm threshold.
Currently a single 500 response is often enough to trigger this alarm.
This means our `create recurring product failed` alarm is almost
always followed by the 5xx alarm. In the spirit of trying to reduce
the number of alarms that we receive to reduce noise and increase the
proportion of meaningful alarms, we decided to up the threshold.