Fix GHAs on dependabot PRs.
Action runs of dependabot PRs don't have access to secrets that are
configured in the repo. This means they can't do any steps that require
authentication like uploading to s3 with riffraff. This change makes it so
all of the regular workflows are skipped on dependabot PRs and introduces a
dependabot specific workflow to that runs the support-frontend & cdk tests.