Add typescript & eslint workflows.
Adds two new workflows: support-frontend-typescript and support-frontend-eslint.
Now that we've got both typescript and lint errors down to 0, we can turn on
these checks in CI to ensure they stay that way.

I've followed the style of dotcom-rendering which is to have them as separate
workflows so that they can run in parallel.