Factor out some hooks from `StripeCardForm.tsx`.
A separate piece of work requires some changes to `StripeCardForm.tsx`.
The file is pretty big and has a mix of stripe/redux/rendering logic going
on which makes it quite difficult to work on. I'm planning to do a few refactors
to chunk the file up a bit and make it easier to work with. This first change
factors out a couple of hooks:

- useSelectedField
- useCardFormFieldStates

The `useSelectedField` hook isn't really needed at all. I just quite like
the aesthetic of having functions that create callbacks to be passed to click
handlers. That is, using:

```ts
onFocus={selectField("CVC")}
```

over:

```ts
onFocus={() => selectField("CVC")}
```

The `useCardFormFieldStates` moves the logic associated with keeping track of the
state of each of the form fields out of main component and into this hook.

I don't anticipate either of these will be reused anywhere else, but really just
using the hooks as a way to move logic out of the component itself.