Remove `handle3DS` from contribs redux state.
This change removes the `handle3DS` function from the contribs
redux state along with the corresponding `setHandleStripe3DS` action.

It was a little tricky to figure out how to remove as it's accessed
inside of a thunk creator & then passed to other thunks (or other
things that are almost thunks). I decided to add it to the
`StripePaymentIntentAuthorisation` so that components that were originally
calling `setHandleStripe3DS` can now just pass a `handle3DS` callback to
the `onPaymentAuthorised` callback prop.