Remove AmazonPay objects from redux store.
This change removes the `AmazonLoginObject` and `AmazonPaymentsObject` from
the redux store, and moves them to component state. We're removing them as
they contain non-serialisable state.

To do this, I've created a `useAmazonPayObjects` hook that is responsible
for initialising the AmazonPay sdk and storing the objects in state. I
initially used this inside of the new `AmazonPayCheckout` component that I
factored out of the `ContributionsSubmit` component. However, that gets
mounted/unmounted as the user changes payment method which makes it a
little awkward as we only want to initialise the sdk once. Instead, I lifted
the `useAmazonPayObjects` hook into the parent `ContributionsSubmit` component
as that is always rendered.