Move recaptcha to the last page of the dd form
This broke the styles somewhat. To avoid having to dig around too much
in the nasty sass, I just moved the error message above the buttons
and simplified the styles. The ErrorMessage component was only being
used by the contributions direct debit form so it's safe to update just
keep the styles we need for this particular use.