Payment API CDK migration phase 1
In this phase, we introduce CDK to the payment-api project. At this stage
we're just using the `CfnInclude` construct to load the existing cloudformation
template. This allows us to get all the plumbing in place (eg generating the
template during CI and pointing riffraff at the new location) without also
generating new versions of all of the resources. In phase 2, we will spin up a
duplicate set of all resources using proper cdk/gu-cdk constructs/patterns.
When we are happy this new set of resources is running as expected, we can
redirect traffic to the new resources. In phase 3, after the old resources are
no longer receiving traffic, we can tear them down.