Payment API CDK migration phase 2.
This phase of the CDK migration involves spinning up a second
set of all of the resources within a stack. This allows us to safely
deploy this change as all of the traffic will be pointed at the existing
infrastructure. We can then test the new infrastructure and when we're happy
everything is working we can point traffic at the new resources.