Add supporter plus support to acquisitions-firehose-transformer (#4447)
This adds support for the new supporter plus product prop. We need to be
able to calculate A/V for the new product. Currently it's a duplication
of the recurring contribution calculation. I've implemented the margin
calculation as a separate method to make it easy to change in future as
we learn more about supporter plus.