Don’t do postcode lookup for long inputs
For certain very long input values, getAddress.io’s postcode lookup
returns a 404 instead of the expected error. To guard against this, this
change avoids calling the API for values too long to be a valid
postcode.

I believe the longest valid postcodes are 8 characters; this check
checks for inputs longer than 10 characters in case I’m mistaken or in
case that changes.