Propagate 4xx errors when getting user type
Before now, if we got a 4xx error from Identity (for example, if the
email address we sent them was invalid), we would return a 500 to the
client.

There are two problems with this. The first is that an invalid email
address here is not a *server-side* error, so a 500 has the wrong
meaning.

The second is that this triggers an alert we have set up to detect 5xx
errors. We call this on the digital subscription checkout page when the
user moves focus out of the email address input, and don’t show the user
the resulting error immediately, so 5xx errors here produce unnecessary
alerts.

This change instead returns the 4xx error we got from Identity to the
frontend, avoiding alerting and returning a more semantically
appropriate error.