Disable payment button if billing id is empty
This is a quick fix rather than a proper solution: since it would be an
error to submit the form if the billing ID was falsy, this change
disables the button in that case.

The proper solution would involve figuring out how we get to a selected
existing payment method that doesn’t have a normal billing ID and fixing
that.