Add request ID to acquisition-events-api logs
It can be hard to tell when looking at these logs which messages are
from which request. The request ID is mentioned in the start-end-report
logs (which I think AWS provides?), but not in our logging in between.
Adding the request ID like this will allow us to filter to a specific
request more easily, hopefully helping to debug issues.