Use correct country to check if state is required
At the moment, if a user is on the UK page but uses Apple Pay and we get
a billing address that’s in the US, we don’t require a non-empty state
on the client-side. We then send the request to the backend, which
fails.

To fix this, this commit uses the same logic for determining the country
when checking whether a state is required as is used when sending the
info to the backend. This way the client-side and server-side validation
should be consistent.