Add stripePaymentMethod to state
At the moment, if a payment via Apple Pay (or presumably other stripe
request button payment methods) fails, the country and state will get
reused for subsequent card payments, potentially causing further
failure. I saw this recently where we got US as the country and no state
from Apple Pay, and when that failed the user retried with card, which
also failed. Retrying with direct debit fixed the problem for this user,
but it shouldn’t be necessary.

Paul Brown fixed this bug for some payment types in [PR
2378](https://github.com/guardian/support-frontend/pull/2378) (in commit
14b6a2b351fdcd5a198d1958b359ad0467daae34) by clearing the country and
state on payment failures. However, since the else branch isn’t taken
when the payment method is Apple Pay, the bug can still happen in that
case.

Rather than deleting the country and state when payment fails, this
change keeps track of whether the current/most recently used payment
method is the payment request button, and only sends the country and
state obtained from the button if it is.

This should fix the bug for Apple Pay users, but I have not verified
that it does – I don’t know how to test it, because Apple Pay won’t let
me add an invalid card, or one that has an invalid billing address, or a
test card. I have confirmed that the happy path still works: using a US
Apple Pay card on the UK page or vice-versa works fine (and the billing
address is set correctly) if the address has a state when needed.

I wanted to set the `stripePaymentMethod` correctly when using Stripe
without the payment request button as well, but it was hard to determine
whether the correct value should be `StripeCheckout` or
`StripeElements`, or a mix of the two. Currently we use a mix: mostly
`StripeCheckout` but with `StripeElements` used once.

I believe the original intended meaning was that `StripeCheckout` meant
Stripe’s low-code
integration (https://stripe.com/docs/payments/checkout), and
`StripeElements` meant the prebuilt UI components that we later switched
to (https://stripe.com/docs/payments/elements). However, we aren’t using
the types that way: as far as I can see we don’t use the low-code
integration at all anymore, but we do still use `StripeCheckout`.

I wanted to clarify the situation (and maybe get rid of one of the two),
but don’t want to spend more time on this. Maybe we can update this in
the future.