Clarify variable name
The StripePaymentRequestButton payment method may mean Apple Pay, but it
also may mean Google Pay (and possibly a third payment method (in
Stripe’s terms, a “wallet”) that I’m unaware of). It’s better for this
variable to be named more specifically in this case, and so I can put
this commit message here.