Allow taking out multiple recurring contributions
It’s been decided to completely remove the restriction on users taking
out multiple recurring contributions. The restriction was causing some
issues, mainly that when a user cancelled their contribution, the
restriction still prevented them from taking out a new one because they
are still listed by parts of the website as a supporter so they can
continue to access the site until the end of their current payment
period.

In the future the restriction may be restored in an updated form (in
conjunction with a dedicated product switching flow), but for now
disabling it is agreed to be the best approach. Stakeholders are aware
that this could potentially lead to people accidentally taking out
multiple contributions, and have accepted the risk.