Store payment method country/state separately
Rather than overwriting the form values, this commit stores the country
and state obtained from a payment request button wallet in a separate
part of the redux state. Then when working out the billing country and
state for validation or for submission to the backend, the code only
uses the payment-method-provided country and state if the current
payment method is the payment request button.

This should fix the following error we’ve encountered: If a user has
invalid country/state info set in their Apple Pay details (e.g. the
country is US but the state is empty), the payment will fail on the
backend. If the user then tries another payment method, the
country/state info from Apple Pay will still be in the redux state, and
will be sent to the backend to fail again.

Another way to fix that issue would be to clear the
payment-method-provided info on error: we already do that for errors
earlier in the process, and we could plausibly do that if the
support-workers execution fails. However, now that we’re doing
the country determination at page load, we would’ve lost that info when
overwriting it with the payment-method-provided info. We could save it
somewhere and restore it, or run the determination function again, but I
prefer this approach. One reason for that is that it puts the choice
logic all in one place: the other approaches require an insertion and
deletion to be co-ordinated in different parts of the code, but this
approach only requires the insertions.

This commit also changes setPayerDetails to reset the user’s email
address from session storage if it’s available, rather than setting to
an empty string. If a payment request button payment fails when the user
was initially signed in, resetting to an empty string would leave them
signed in, but clear their email address from the form, leaving an empty
disabled input. They’d still be able to sign out (by clicking the “Not
you?” link) and then re-enter their email address, but it’s a bit of an
odd experience. Getting the email address from session storage should
handle this case. (I still can’t test this because I can’t produce a
failing payment request button payment.)