Narrow IsoCountry type to union of literals
The IsoCountry type is currently inferred as `string`, rather than the
more specific union of string literals that it looks like was intended.
From talking to other developers, I think this is leftover from the
switch from flow to typescript, and that the type was correctly inferred
with flow.

Using an array of the possible values like this allows it to be used
with zod as a zod enum – I initially tried a straightforward union type
but that didn’t work with it.

I’ve tested locally that the supporter plus checkout, newspaper
subscription checkout, and guardian weekly checkout all function
correctly for a variety of countries and send valid requests to the
backend.