Replace BuildInfoKey.constant with (_ -> _)
BuildInfoKey.constant here is an implicit, the name of which was changed
in [this commit in
2020](https://github.com/sbt/sbt-buildinfo/blob/61539b1b0452f3d24553d0fe2f2bd8ebdd9f914b/src/main/scala/sbtbuildinfo/package.scala#L9).
The change wasn’t included in the changelog, I’m guessing because they
assumed no-one would be using the implicit directly. (This took a while
to find, so I’m a bit annoyed about it.)

Using a tuple and allowing the implicit conversion to be used should
prevent this erroring again the next time they change the name.