Replace .enumEntry access with .in() call
The `enumEntry` val has been made private in [PR
343](https://github.com/lloydmeta/enumeratum/pull/343), which means we
need to do this access differently. The `.in()` method assembles a list
from the parameters and checks if the enum value is contained in the
list, so this should be equivalent. It feels like this didn’t need to be
changed.