Try updating java version to 11
All apps [should move from java 8 to java
11](https://docs.google.com/document/d/1ZR-YnaXCT5_gLVmTCeGs0mWd3KPaAozPjQK8uUzHZ9w/edit?pli=1#),
but this one is still on java 8. Some recent dependency upgrades like
scaffeine are starting to require java 11 as well, so sticking with java
8 would mean sticking with some old dependency versions.

I have updated the github action for the CI, as well as which amigo
recipe is used by riff raff. There’s also a debian package specified in
the build.sbt: I’m not sure what this is used for, as I would’ve
expected the JRE to be provided by amigo.