Document uncertainty about sbt dependencyOverrides
I’m a little nervous about our use of dependencyOverrides, and whether
it’s definitely safe for us to do. However, we do it in a few places: it
seems to be a standard practice of ours. I wanted to document this here
following a discussion with a senior engineer who shared my concerns:
to record that we’re uncertain about it, but it has worked fine so far.