Don’t guess state from geolocation data
At the moment, we guess a user’s state on page load based on geolocation
data provided by Fastly. This allows users to avoid needing to fill out
the state field on the form.

However, the guessed state is wrong relatively often, and has been
causing us trouble. For example, a US user may have a
geolocation-guessed state that is in Canada (presumably if they are near
the border), and then our client-side validation needs to check that the
current state value fits with the current country value.

Since it doesn’t seem like too much effort to ask the user to go to to
enter their state along with the other info we ask them for, this commit
stops the site guessing the user’s state from Fastly’s geolocation data.