Print debug info earlier
Recently, I’ve noticed that some errors mean we don’t see the nicely
formatted debug info. I think this is because the errors happen before
we get to the line that’s currently doing that logging. Moving that
earlier in the execution flow should make our debugging a bit nicer.