Add endpoints for calling PaperRound’s API
To expand paper delivery outside the M25 using PaperRound, we’ll need to
offer users a choice of delivery agents for their postcode. This change
makes a prototype API on the support-frontend backend which calls
PaperRound’s test API to check the agents available for a certain
postcode. (I’ve also added the other API endpoints to the service while
here.)

- Model PaperRound API and write JSON decoders

  Fixtures taken from responses from the test API

- Write separate types for support-frontend

  I wanted to change the structure received by the frontend, so I’ve
  used separate types for the API called by the frontend. I didn’t want
  to have decoders and encoders for the PaperRound types that don’t
  round-trip. It also seems a good idea to isolate the dependency on
  PaperRound’s types from the frontend, but I’m less confident about
  that one.

- Write custom json encoding for frontend

  The derived encoding of the `GetAgentsResponse` ADT doesn’t seem very
  nice to handle from TypeScript, so I’ve written a custom one that I
  think will be nicer to consume from TypeScript.