acquisition-events-api: Disable 5xx monitoring
At the moment, the lambda that sends events to BigQuery frequently takes
more than 30 seconds to run, which is the timeout limit for API Gateway:
this causes the API Gateway to give up on the lambda and return a 5xx to
the client, even though the lambda then succeeds shortly afterwards.

Since we can’t quickly fix the lambda to run in a shorter amount of
time, this alarm isn’t worth having: it generates a really high level
of (approximately) false positives.

Therefore this change disables the alarm for now, until we either speed
up the lambda or replace it (possibly with the proposed new architecture
where the API Gateway is connected directly to an event bus).