Point prout to subscribe page
The /contribute page now does an oauth redirect, which prout seems to be
having trouble with. ([The
logs](https://prout-bot.herokuapp.com/view/guardian/support-frontend)
report `Could not read from this url, got java.net.ProtocolException:
Too many follow-up requests: 21`.)

We think this is because the oauth redirects require cookies to be set
in order to function correctly, while Prout doesn’t set cookies.

Handily, the /subscribe endpoint does not (currently) require an oauth
redirect, and so this should fix the problem.