BigQueryService: close resources
In commit fc82dde5c51f863f7c8d1c479592204ba4390389, I begin closing
resources after each request. However, while the acquisition-events-api
can handle this, the payment-api can’t! The former creates a new service
on every request, but the latter doesn’t, so on the first request the
client was closed and never reopened, causing single contributions to
break. (These changes were reverted in commit
6dbcd65e785a0498adbe9d0cc709d6a6fa8760a4).

This commit moves the closing logic so the payment-api doesn’t close the
big query client after the first request, while the
acquisition-events-api can still close and re-open on every request.

Another option would’ve been to create a new service on every request to
the payment api, but the storage write api seems designed to be used in
more of a long-running server context like this.

I also forgot to include support-workers, so I’ve done that now. We
haven’t seen any problems with that though, possibly because of how the
underlying lambdas are handled for stepfunctions?