Synchronise deliveryAgent on frontend and backend
These were out of sync, which I hadn’t noticed: the delivery agent
returned from PaperRound’s API is a number, whereas the backend expected
it from the frontend as a string – I’ve changed this to be a number
everywhere.

Also, the backend expects it in the product but the frontend was sending
it at the top level of the create request – I’ve moved this to the
product.