payment-api: Add optional postal code field
We want to [collect zip codes for US
contributions](https://trello.com/c/VSe61QY5/1436-support-checkout-back-end-changes-card-move-zip-code-out-of-the-card-payment-section-across-both-single-recurring-in-us-and-make),
and this change adds an optional postal code field to hold these values.

I’ve gone with “postal code” rather than “zip code” because even though
we’re only storing it for US contributions at the moment, we may expand
that in future, and “postal code” appears to be the more general term,
judging by [the Wikipedia
page](https://en.wikipedia.org/wiki/Postal_code).