DO_NOT_MERGE: Hardcode some test agents
PaperRound’s test API is currently not returning any agents for the
postcodes we’ve been using. So that we can keep testing in CODE, this
commit hardcodes responses for those postcodes.

I’ve also included a larger number of test agents, so we can see how the
frontend handles that.