National Delivery: thank you email changes
- Use new extension name

Added to membership-workflow in
https://github.com/guardian/membership-workflow/pull/500, which has been
merged first.

- Include delivery agent details

Only the name is needed so far, but I might as well include the others
while I’m here.

- Use test PaperRound API for test users in PROD

At the moment we’re pointed at the test API in PROD anyway, so the
change isn’t actually visible.

- Update config for new provider setup

Both the parameter store config and the support-workers S3 config.

- Use wrapper type for agent ID

Also move most of the paperround type definitions to support-models, so
they can be imported to all the right places. (Move the docs as well, to
keep them in the same place as the type definitions.)