AgentsEndpoint.Response: Remove message field
This field has been removed by PaperRound from the API, presumably
because they didn’t intend for us to rely on it. We weren’t actually
_using_ the field, so we shouldn’t have been parsing it in the first
place!

At the moment this is breaking the sending of thank you emails from
support-workers, so this commit should fix that by no longer parsing it.