fix(identity): change logic of `GU_SO` cookie use
We discovered an issue in the way that the GU_SO cookie works, meaning that a user could sign out, then immediately sign in again as a new user, but the tokens for the old user had not been cleared yet.
The problem and solution have been outlined in detail within this PR https://github.com/guardian/gateway/pull/2529
And the documentation was updated in this PR https://github.com/guardian/gateway/pull/2530

Essentially the fix is to read the `GU_SO` cookie and get the timestamp for when a user signs out, and compare it to the `iat` claim of the ID token. This leads to one of three scenarios:
1. If the `GU_SO` value is > `iat` then the user has signed out recently, so the tokens are invalid, get new tokens if needed
2. If the `GU_SO` value is < `iat` then the user has not signed out recently, so the tokens are valid, continue
3. If the `GU_SO` cookie is missing then the user has not signed out recently, so the tokens are valid, continue