Migrate securePadlock and securePadlockCircle SVGs to React components (#6479)
When the div was added as a container for the svg element, a little extra
height was added resulting in chromatic diffs. Making the svg display block
seems to solve this: https://stackoverflow.com/a/10426847.

The colour is now controlled by a prop of new svg padlock components. This
feels more intuitive than controlling via css with fill: currentColour (which
wasn't working correctly with the additional div).