Remove overlay if one time card payment fails server side
Previously we noticed that if there was a server side failure on the one
time checkout, you'd see an error message but the "processing payment"
overlay would remain up, which meant you couldn't interact with the form
to fix whatever the problem was.

We've also added some playwright coverage for this scenario.