Extend client side input validation to reject all emojis (#6661)
We had a case in prod where a flag emoji made it to the backend (and was
correctly rejected). However we expected this to be rejected in the
client side validation. It turns out that flag emojis don't match
Extended_Pictographic so the regex which said "anything other than
Extended_Pictographic" let it through. We now reject
Extended_Pictographic and Emoji_Presentation.