Add min amount validation to server side one time payments (#6665)
We already have client side validation for this, but we'd like to enforce
server side too as this makes us more robust.

I've tested this in CODE. In order to do this I had to temporarily remove the
client side validation. The error message we show in this case is quite
generic, but I think in general we don't expect users to hit this as the client
side validation should kick in before the request even makes it to the backend.