Bump support-service-lambas package to bring in Guardian Ad-Lite (#6692)
In this PR we update the support-service-lambdas package and handle these
changes in support-frontend. This includes handling new products (including
GuardianAdLite) added to the product catalog since the last package bump. We
also leverage that we can now distinguish between active and inactive products
at the type level (this reduces the burden for handling newly added products,
as we only have to handle products we actually sell).

There are a couple of places where we've had to handle GuardianPatron and
OneTimeContribution, as they've been added to the product catalog as active
products, even though they aren't sold via the generic checkout.  So our
abstraction here is slightly imperfect, but I think it's close enough and
adding a new concept to the product catalog to represent this feels out of
scope and too support-frontend specific.

Note: this PR only covers the client-side changes and not the backend changes.
This means that you can access the generic checkout for Guardian Ad-Lite but
can't put through a transaction yet - an error is returned by the backend. This
work will be completed in a future PR.