Guardian Ad-Lite backend rename (#6700)
This PR updates the backend and support-workers code to handle the new product
Guardian Ad-Lite. I've done this as a rename, so after this PR is merged
GuardianLight purchases will no longer work (even though the client will
continue handling GuardianLight until it's removed). The playwright tests have
been updated.

Changes which have an impact outside of this codebase:

* The cookie dropped has been renamed - communicate this to affected teams.
* The email IDs have changed, so will need updating in membership-workflow
  (Update Ad-Lite thank you email IDs to reflect new name membership-workflow#572
  - should be deployed at the same time).
* The acquisition ID has changed - communicate this to affected teams.

Now that the backend supports GuardianAdLite (and no longer supports
GuardianLight) I've moved over some final references in the frontend to use the
new product name, e.g. we now link from the landing page to the checkout with
the new product ID.