Prevent Guardian Ad-Lite purchase by users who have either of `adFree` or `allowRejectAll ` benefits already (#6675)
If a non signed-in user already has the adFree or allowRejectAll benefit, they
shouldn't be able to purchase Guardian Ad-Lite. We're assuming that a signed-in
user with adFree or allowRejectAll wouldn't make it to the checkout as there's
no user journey for this. It doesn't make sense to buy Guardian Ad-Lite if you
get ad free reading or can already reject all.

This is configured as a touchpoint service, similar to other examples. This is
to support test users in PROD, which should talk to the CODE user benefits API.